+++ Runtime/Tweens/Components/TweenDestroyCommand.cs
@@ -1,0 +1,14 @@
+﻿using Unity.Entities;
+
+namespace Timespawn.EntityTween.Tweens
+{
+    public struct TweenDestroyCommand : IBufferElementData
+    {
+        public TweenDestroyCommand(int id)
+        {
+            Id = id;
+        }
+
+        public int Id;
+    }
+}